Refactor config to match with the latest docs
