--- .storybook/addons.js
@@ -1,4 +1,0 @@
-import '@storybook/addon-notes/register';
-import '@storybook/addon-knobs/register';
-import '@storybook/addon-a11y/register';
-import '@storybook/addon-viewport/register';
