--- .storybook/stories.pattern
@@ -1,1 +1,0 @@
-../packages/*/*/*/*.stories.jsx
